Move crypto related files to one folder
