+++ src/framework/Network/MangosSocketImpl.h
@@ -16,8 +16,8 @@
 #include "SharedDefines.h"
 #include "ByteBuffer.h"
 #include "Database/DatabaseEnv.h"
-#include "Auth/BigNumber.h"
-#include "Auth/Sha1.h"
+#include "Crypto/BigNumber.h"
+#include "Crypto/Hash/Sha1.h"
 #include "WorldSession.h"
 #include "Log.h"
 #include "DBCStores.h"
