+++ src/game/Anticheat/WardenAnticheat/WardenWin.cpp
@@ -32,8 +32,8 @@
 #include "Unit.h"
 #include "Chat.h"
 #include "WorldSession.h"
-#include "Auth/BigNumber.h"
-#include "Auth/HMACSHA1.h"
+#include "Crypto/BigNumber.h"
+#include "Crypto/Hash/HMACSHA1.h"
 #include "ByteBuffer.h"
 #include "Database/DatabaseEnv.h"
 #include "Player.h"
