+++ src/game/Handlers/CharacterHandler.cpp
@@ -31,11 +31,9 @@
 #include "Player.h"
 #include "Guild.h"
 #include "GuildMgr.h"
-#include "Auth/md5.h"
 #include "ObjectAccessor.h"
 #include "Group.h"
 #include "Database/DatabaseImpl.h"
-#include "PlayerDump.h"
 #include "SocialMgr.h"
 #include "Util.h"
 #include "Language.h"
