+++ src/realmd/AuthSocket.cpp
@@ -24,7 +24,7 @@
 */
 
 #include "Common.h"
-#include "Auth/Hmac.h"
+#include "Crypto/Hash/Hmac.h"
 #include "Auth/base32.h"
 #include "Database/DatabaseEnv.h"
 #include "Config/Config.h"
