+++ src/shared/Crypto/Authentication/SRP6.cpp
@@ -18,7 +18,7 @@
 
 #include "Common.h"
 #include "Log.h"
-#include "Auth/HMACSHA1.h"
+#include "Crypto/Hash/HMACSHA1.h"
 #include "Auth/base32.h"
 #include "SRP6.h"
