+++ src/shared/Crypto/Authentication/SRP6.h
@@ -20,8 +20,8 @@
 #define _SRP6_H
 
 #include "Common.h"
-#include "Auth/BigNumber.h"
-#include "Auth/Sha1.h"
+#include "Crypto/BigNumber.h"
+#include "Crypto/Hash/Sha1.h"
 #include "ByteBuffer.h"
 
 #define HMAC_RES_SIZE 20
