+++ src/shared/Crypto/Hash/Hmac.cpp
@@ -16,8 +16,8 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#include "Auth/Hmac.h"
-#include "BigNumber.h"
+#include "Hmac.h"
+#include "../BigNumber.h"
 
 HmacHash::HmacHash(uint8 const* data, int length)
 {
